paginate job offers, refactored will_paginate custom renderer
