--- app/views/job_offers/_category.html.haml
@@ -1,26 +1,0 @@
-- if category.depth == 0
-  - sub_categories = @categories.select{|c| c.parent_id == category.id}
-  - sub_sub_categories = @categories.select{|c| sub_categories.map(&:id).include?(c.parent_id) }
-  - concerned_categories = [category] + sub_categories + sub_sub_categories
-  - job_offers = @job_offers.select{|x| concerned_categories.map(&:id).include?(x.category_id) }
-  - direct_job_offers = @job_offers.select{|x| category.id == x.category_id }
-  - if job_offers.any?
-    %h2= category.name
-    - if direct_job_offers.any?
-      = render partial: 'job_offer', collection: direct_job_offers, locals: {category: category}
-    = render partial: 'category', collection: @categories.select{|x| x.parent_id == category.id}
-- elsif category.depth == 1
-  - sub_categories = @categories.select{|c| c.parent_id == category.id}
-  - sub_sub_categories = @categories.select{|c| sub_categories.map(&:id).include?(c.parent_id) }
-  - concerned_categories = [category] + sub_categories + sub_sub_categories
-  - job_offers = @job_offers.select{|x| concerned_categories.map(&:id).include?(x.category_id) }
-  - if job_offers.any?
-    - direct_job_offers = @job_offers.select{|x| category.id == x.category_id }
-    %h3= category.name
-    - if direct_job_offers.any?
-      = render partial: 'job_offer', collection: direct_job_offers, locals: {category: category}
-    = render partial: 'category', collection: @categories.select{|x| x.parent_id == category.id}
-- elsif category.depth == 2
-  - job_offers = @job_offers.select{|x| x.category_id == category.id }
-  - if job_offers.any?
-    = render partial: 'job_offer', collection: job_offers, locals: {category: category}
