+++ app/views/job_offers/_job_offer.html.haml
@@ -1,4 +1,5 @@
-- cache [category, job_offer] do
+- cache [job_offer] do
+  - category = job_offer.category
   .rf-pb-2w.rf-pt-2w
     = link_to job_offer do
       .d-flex.align-items-center
