+++ app/views/job_offers/index.html.haml
@@ -4,4 +4,6 @@
     .rf-col-xs-12.rf-col-md-3
       = render partial: 'filters'
     .rf-col-xs-12.rf-col-md-9
-      = render partial: 'job_offers_list'
+      = t('.job_offers_count', count: @job_offers.size)
+      = render partial: 'job_offer', collection: @job_offers
+      = will_paginate @job_offers, class: 'justify-content-center rf-mt-3w', renderer: RfRenderer
