+++ config/application.rb
@@ -63,6 +63,8 @@ module Civilsdeladefense
     end
 
     config.eager_load_paths << Rails.root.join('lib/services')
+    config.eager_load_paths << Rails.root.join('lib/renderers')
     config.autoload_paths << Rails.root.join('lib/services')
+    config.autoload_paths << Rails.root.join('lib/renderers')
   end
 end
