+++ config/locales/fr.yml
@@ -109,17 +109,16 @@ fr:
   job_offers:
     index:
       title: "Nos offres"
+      job_offers_count:
+        zero: "Aucune offre ne correspond à votre recherche"
+        one: "1 offre d'emploi trouvée"
+        other: "%{count} offres d'emploi trouvées"
     filters:
       select_category: "Secteur"
       blank_category: "Tous les secteurs"
       select_contract: "Contrat"
       blank_contract: "Tous les contrats"
       search: "Rechercher"
-    job_offers_list:
-      job_offers_count:
-        zero: "Aucune offre ne correspond à votre recherche"
-        one: "1 offre d'emploi trouvée"
-        other: "%{count} offres d'emploi trouvées"
     job_offer:
       show: "Voir le détail de l'offre"
       published_date: 'Publiée il y a %{time_ago_in_words}'
